Moved getExperimentBranch to /utils, extended it to both home and scan results files, added UTMs to /scan page
