+++ scan-results.js
@@ -5,11 +5,27 @@ const { URL } = require("url");
 const HIBP = require("./hibp");
 const sha1 = require("./sha1-utils");
 
+const { getExperimentBranch } = require("./controllers/utils");
+
+const AppConstants = require("./app-constants");
+const EXPERIMENTS_ENABLED = (AppConstants.EXPERIMENT_ACTIVE === "1");
 
 const scanResult = async(req, selfScan=false) => {
 
   const allBreaches = req.app.locals.breaches;
   let scannedEmail = null;
+
+  let experimentBranch = null;
+  let isUserInExperiment = null;
+  let experimentBranchB = null;
+
+  if (EXPERIMENTS_ENABLED) {
+    const coinFlipNumber = Math.random() * 100;
+    experimentBranch = getExperimentBranch(req, coinFlipNumber);
+    req.session.experimentBranch = experimentBranch;
+    isUserInExperiment = (experimentBranch === "vb");
+    experimentBranchB = (experimentBranch === "vb" && isUserInExperiment);
+  }
 
   const title = req.fluentFormat("scan-title");
   let foundBreaches = [];
@@ -98,6 +114,9 @@ const scanResult = async(req, selfScan=false) => {
     fullReport,
     userDash,
     scannedEmailId,
+    experimentBranch,
+    isUserInExperiment,
+    experimentBranchB,
   };
 };
